JS API/Document/Hooks: Fix - Relocate to full command path.
