+++ assets/dev/js/editor/document/hooks/data/document/elements/delete/index.js
