+++ assets/dev/js/editor/document/hooks/data/document/elements/delete/section-columns.js
