+++ assets/dev/js/editor/document/hooks/data/document/elements/paste/index.js
