+++ assets/dev/js/editor/document/hooks/data/document/elements/paste/is-paste-enabled.js
