+++ assets/dev/js/editor/document/hooks/data/document/elements/settings/handle-dynamic.js
