+++ assets/dev/js/editor/document/hooks/data/document/elements/settings/resize-column.js
