+++ assets/dev/js/editor/document/hooks/data/document/elements/settings/set-structure.js
