+++ assets/dev/js/editor/document/hooks/data/index.js
@@ -1,7 +1,7 @@
 // Alphabetical order.
 
-export * from './create/';
-export * from './delete/';
-export * from './move/';
-export * from './paste/';
-export * from './settings/';
+export * from './document/elements/create/';
+export * from './document/elements/delete/';
+export * from './document/elements/move/';
+export * from './document/elements/paste/';
+export * from './document/elements/settings/';
